Use actual no person image in test to show its failing
The person_detection_test was only testing that the model worked for classifying a person.

It said it was testing classifying a non person image but in reality it was just zeroing the input tensors.

Now that I have it considering the no person image it fails to classify it properly.