Add Support for building Micropython using project generator (#702)
Micropython needs the C++ source files to have .cpp file extensions
where as Tensorflow uses the .cc extension.

Adds a new flag, --rename-cc-to-cpp,  to the example generator script that will cause the copied
.cc files to become .cpp files in the destination directory.

Apply the .cc to .cpp rename for the examples aswell.

Add new test to the project generation tests.